feat: make `KiwiDrive` an API
